+++ src/main/java/io/github/sunday/devfolio/DTO/auth/UserSignupDto.java
@@ -1,4 +1,4 @@
-package io.github.sunday.devfolio.dto;
+package io.github.sunday.devfolio.dto.auth;
 
 import lombok.Getter;
 import lombok.Setter;
