+++ src/main/java/io/github/sunday/devfolio/DTO/portfolio/PortfolioSearchRequestDto.java
@@ -1,7 +1,7 @@
 package io.github.sunday.devfolio.dto.portfolio;
 
 import io.github.sunday.devfolio.annotation.portfolio.PortfolioCategoryValid;
-import io.github.sunday.devfolio.enums.PortfolioSort;
+import io.github.sunday.devfolio.enums.portfolio.PortfolioSort;
 import jakarta.validation.constraints.AssertTrue;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
