+++ src/main/java/io/github/sunday/devfolio/repository/auth/EmailVerificationRepository.java
@@ -1,4 +1,4 @@
-package io.github.sunday.devfolio.repository;
+package io.github.sunday.devfolio.repository.auth;
 
 import io.github.sunday.devfolio.entity.table.user.EmailVerification;
 import org.springframework.data.jpa.repository.JpaRepository;
