+++ src/main/java/io/github/sunday/devfolio/repository/user/UserRepository.java
@@ -1,4 +1,4 @@
-package io.github.sunday.devfolio.repository;
+package io.github.sunday.devfolio.repository.user;
 
 import io.github.sunday.devfolio.entity.table.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
