+++ src/main/java/io/github/sunday/devfolio/service/auth/EmailService.java
@@ -1,4 +1,4 @@
-package io.github.sunday.devfolio.service;
+package io.github.sunday.devfolio.service.auth;
 
 public interface EmailService {
     void sendVerificationCode(String email, String verificationCode);
