+++ src/main/java/io/github/sunday/devfolio/service/common/SecureImageService.java
@@ -2,7 +2,7 @@ package io.github.sunday.devfolio.service.common;
 
 import io.github.sunday.devfolio.dto.common.ImageUploadResult;
 import io.github.sunday.devfolio.dto.common.MultipartImage;
-import io.github.sunday.devfolio.enums.ImageTarget;
+import io.github.sunday.devfolio.enums.common.ImageTarget;
 import io.github.sunday.devfolio.exception.common.ImageUploadException;
 import io.github.sunday.devfolio.exception.common.ImageValidationException;
 import io.github.sunday.devfolio.exception.common.IncorrectImageTargetException;
